// src/shared/types/api/offers.types.generated.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/api/offers": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Create offer */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody: {
                content: {
                    "application/json": components["schemas"]["OfferPostBody"];
                };
            };
            responses: {
                /** @description Created */
                201: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["OfferId"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        OfferId: {
            /** @description Id of the offer */
            id?: string;
        };
        OfferPostBody: {
            /** @description Id of the advertiser */
            advertiser_id: number;
            /** @description Model alias */
            model: string;
            /** @description Full offer name */
            name: string;
            /** @description Short offer name */
            short_name: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type OfferId = components['schemas']['OfferId'];
export type OfferPostBody = components['schemas']['OfferPostBody'];
export type $defs = Record<string, never>;
export type operations = Record<string, never>;
export enum ApiPaths {
    PostApiOffers = "/api/offers"
}